refactor: isolate multipart filename security coverage
